Updated point vector indexed to be long long. Using integer could result in potential loss of data.

+++ pyvoronoi/voronoi.cpp
@@ -29,8 +29,8 @@ std::vector<Segment> VoronoiDiagram::GetSegments() {
 
 
 
-std::vector<int> VoronoiDiagram::GetIntersectingSegments(){
-    std::vector<int> overlapping_indexes;
+std::vector<long long> VoronoiDiagram::GetIntersectingSegments(){
+    std::vector<long long> overlapping_indexes;
     for (auto it_left = segments.begin(); it_left != segments.end(); ++it_left) {
         Segment* segmentLeft = &(*it_left);
 
@@ -47,8 +47,8 @@ std::vector<int> VoronoiDiagram::GetIntersectingSegments(){
     return overlapping_indexes;
 }
 
-std::vector<int> VoronoiDiagram::GetDegenerateSegments(){
-    std::vector<int> degenerate_indexes;
+std::vector<long long> VoronoiDiagram::GetDegenerateSegments(){
+    std::vector<long long> degenerate_indexes;
     for (auto it_left = segments.begin(); it_left != segments.end(); ++it_left) {
         Segment* segment = &(*it_left);
         if (segment->p0.X == segment->p1.X && segment->p0.Y == segment->p1.Y)
@@ -57,8 +57,8 @@ std::vector<int> VoronoiDiagram::GetDegenerateSegments(){
     return degenerate_indexes;
 }
 
-std::vector<int> VoronoiDiagram::GetPointsOnSegments(){
-    std::vector<int> degenerate_indexes;
+std::vector<long long> VoronoiDiagram::GetPointsOnSegments(){
+    std::vector<long long> degenerate_indexes;
     for (auto it_point = points.begin(); it_point != points.end(); ++it_point) {
         Point* point = &(*it_point);
         for (auto it_segment = segments.begin(); it_segment != segments.end(); ++it_segment) {
